Fix a problem where requireCaptcha() does not display as required
The 'captcha required' problem is cause by the shorthand we use to call
Captcha.

--
requireCaptcha('Are you human?')
--

The above snippet does not work because the 'require<item>()' shorthand is
expecting at least two parameters, and it is skipping the first '$name'
parameter. (later in the code this is allowed)

This commit allows the code above to work as expected.